+++ src/set.cpp
@@ -57,7 +57,7 @@ dplyr::BoolResult compatible_data_frame_nonames(DataFrame x, DataFrame y, bool c
   if (convert) {
     for (int i=0; i<n; i++) {
       try {
-        boost::scoped_ptr<JoinVisitor> v(join_visitor(x[i], y[i], "x", "x", true, true));
+        boost::scoped_ptr<JoinVisitor> v(join_visitor(x[i], y[i], SymbolString("x"), SymbolString("x"), true, true));
       } catch (...) {
         return no_because("incompatible");
       }
@@ -118,23 +118,22 @@ dplyr::BoolResult compatible_data_frame(DataFrame x, DataFrame y, bool ignore_co
   CharacterVector why;
   if (names_y_not_in_x.size()) {
     std::stringstream ss;
-    ss << "Cols in y but not x: " << collapse(names_y_not_in_x) << ". ";
-    why.push_back(ss.str());
+    ss << "Cols in y but not x: " << collapse_utf8(names_y_not_in_x) << ". ";
+    why.push_back(String(ss.str(), CE_UTF8));
   }
 
   if (names_x_not_in_y.size()) {
     std::stringstream ss;
-    ss << "Cols in x but not y: " << collapse(names_x_not_in_y) << ". ";
-    why.push_back(ss.str());
+    ss << "Cols in x but not y: " << collapse_utf8(names_x_not_in_y) << ". ";
+    why.push_back(String(ss.str(), CE_UTF8));
   }
 
   if (why.length() > 0) return no_because(why);
 
   IntegerVector orders = r_match(names_x, names_y);
 
-  String name;
   for (int i=0; i<n; i++) {
-    name = names_x[i];
+    SymbolString name = names_x[i];
     SEXP xi = x[i], yi = y[orders[i]-1];
     boost::scoped_ptr<SubsetVectorVisitor> vx(subset_visitor(xi));
     boost::scoped_ptr<SubsetVectorVisitor> vy(subset_visitor(yi));
@@ -147,12 +146,12 @@ dplyr::BoolResult compatible_data_frame(DataFrame x, DataFrame y, bool ignore_co
     if (!compatible) {
       if (ss.str() == "") {
         ss << "Incompatible type for column '"
-           << name.get_cstring()
+           << name.get_utf8_cstring()
            << "': x " << vx->get_r_type()
            << ", y " << vy->get_r_type();
       }
 
-      why.push_back(ss.str());
+      why.push_back(String(ss.str(), CE_UTF8));
     }
 
   }
@@ -220,7 +219,7 @@ dplyr::BoolResult equal_data_frame(DataFrame x, DataFrame y, bool ignore_col_ord
     if (! track_y.empty()) ss << track_y.str() << ". ";
     if (! track_mismatch.empty()) ss << track_mismatch.str();
 
-    return no_because(ss.str());
+    return no_because(CharacterVector::create(String(ss.str(), CE_UTF8)));
   }
 
   if (ok && ignore_row_order) return yes();